+++ services/backend/factory-service/src/main/java/com/veteroch4k/factory_service/repository/OperationRepository.java
@@ -3,6 +3,6 @@ package com.veteroch4k.factory_service.repository;
 import com.veteroch4k.factory_service.models.Operation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OperationRepository extends JpaRepository<Operation, Long> {
+public interface OperationRepository extends JpaRepository<Operation, Integer> {
 
 }
